introduce scheduler + docs
