+++ docker-compose.yml
@@ -13,8 +13,8 @@ services:
       - ./docker/postgres/init-scripts/db-setup.sh:/docker-entrypoint-initdb.d/db-setup.sh
       - ./mount/storage:/host_os/mount:rw
       
-  cvebox:
-    container_name: cvebox
+  cvebox-engine:
+    container_name: cvebox-engine
     image: cvebox:local
     restart: unless-stopped
     environment: 
@@ -28,6 +28,6 @@ services:
       - ./dist:/cvebox:rw
       - ./node_modules:/cvebox/node_modules:rw
       - ~/.aws:/home/archivebox/.aws:rw
-    command: nodemon /cvebox/main.js
+    command: nodemon /cvebox/main.js engine
     # stdin_open: true
     # tty: true
