+++ docs/engine.md
@@ -1,0 +1,79 @@
+# Archiver Engine 
+
+The actual Archiving consists of a small NodeJS application that is layered on the official ArchiveBox Docker Container. 
+
+
+## Overview
+
+The worker utilizes two external services:
+1. A Postgres database (`DB`) to represent the logical state of all archived assets.
+2. Block storage on AWS `S3` storage.
+
+The general principals are:
+- All assets are stored in `S3`
+- Assets can be publicly accessed (either directly or via a CloudFront CDN) with no server side processing. 
+- The `DB` can be used to track archiving operations, provide troubleshooting, and enable search + other advanced functionality. 
+
+
+
+## The Process (Simplified)
+
+In it's current state, the process can be summarized as:
+- Reads requested CVEs to archive from mounted file
+  - Currently `/etc/cvebox/cvebox.config.json`
+  - Finalized more likely to be parameterized / env var / poll queue / etc
+- Processes CVE(s) 1-by-1
+  - Fetch `References` from CVE api
+  - Run `ArchiveBox` for references
+  - Process `Archive`
+    - Insert logical `Asset`s in database
+    - Upload `jpg` + `warc` to `S3`
+
+
+## Terminology
+
+- `Main` - the main entrypoint for the node app. This will delegate between `engine` and `scheduler` behavior. 
+- `Archive` - the assets produced by `ArchiveBox`
+- `ArchiveModel` - our logical interpretation of an Archive (in the temp directory)
+- `CVE` - the CVE. 
+- `Reference` - a URL reference from the model of the CVE
+
+## Detailed Execution Flowchart
+---
+```mermaid
+---
+title: CveBox Execution Overview
+---
+flowchart TD
+    A[Load config] -->|/etc/cvebox/cvebox.config.json| A1(Partition CVE)
+    A1 --> A2[Partition CVE]
+    A2 --> A3[Resolve References]
+    A4@{ shape: brace-l, label: "Filters Recently Downloaded, Excluded, or any 'Do Not Download'",  } --> A3
+    A3 --> C[Prep TempDir]
+    C --> C1[Initialize Run]
+    C1 --> D[Partition Batch]
+    D --> E[Invoke AB on Partition P]
+    E --> F[Parse Archives & Assets]
+    F --> G[Upsert Domain + Resource]
+    G --> H[Insert Run and RunAsset]
+    H --> I[Upload to S3]
+    I --> J{Did Upload Succeed}
+    J -->|Yes| K[Mark RunAsset available]
+    J -->|No| L[Mark RunAsset Failed]
+    K --> M{Did Upload Succeed}
+    L --> M{More Archives to Process?}
+    M --> |Yes| F
+    M --> |No| N{More Partitions to Process}
+    N --> |Yes| E
+    N --> |No| O[Finish Run]
+    O --> P{More CVEs?}
+    P --> |Yes| A2
+    P --> |No| P1[Finalize Details]
+
+    %% B --> C{Let me think}
+    %% C -->|One| D[Laptop]
+    %% C -->|Two| E[iPhone]
+    %% C -->|Three| F[fa:fa-car Car]
+```
+
+TODO: make a proper Sequence diagram.
