+++ docs/overview.md
@@ -1,0 +1,10 @@
+# Archiver Architecture
+
+The archiver is comprised of two components: 
+- `scheduler` - enables queueing of CVEs to potentially archive.
+- `engine` - the actual application that consumes from the queue and archives CVE references.
+
+
+# Project Setup
+
+The only assumption this project makes is that the runtime environment will have AWS access (either via Environment Variable credentials or implicit IAM roles when deployed).
