+++ docs/scheduler.md
@@ -1,0 +1,3 @@
+# Archiver Scheduler
+
+A stub for a potential API to queuing work to be scheduled.
