+++ package.json
@@ -5,12 +5,12 @@
   "main": "index.js",
   "scripts": {
     "dev": "run-p dev:cvebox dev:watch:ts dev:watch:sql",
-    "dev:cvebox": "docker compose up cvebox",
+    "dev:cvebox": "docker compose up cvebox-engine",
     "dev:watch:ts": "npx tsc --watch",
     "dev:watch:sql": "bash scripts/watch.sh",
     "docker:build": "docker build -t cvebox:local .",
     "docker:bash": "npm run docker:run bash",
-    "docker:run": "docker compose exec -it cvebox"
+    "docker:run": "docker compose exec -it cvebox-engine"
   },
   "keywords": [ ],
   "author": "",
