+++ src/engine.ts
@@ -1,0 +1,32 @@
+import { readFile } from "node:fs/promises"
+import { processCVE } from "./cve/processor"
+
+interface CveBoxConfig {
+  cves: string[]
+}
+
+async function loadConfig() {
+  const f = process.env.CVEBOX_CONFIG_PATH || '/etc/cvebox/cvebox.config.json'
+  try {
+    const raw = await readFile(f, 'utf8')
+    const config: CveBoxConfig = JSON.parse(raw)
+    return config
+  } catch (err: any) {
+    if (err.code == 'ENOENT') {
+      throw new Error('Could not load config file: ' + f)
+    }
+    throw err
+  }
+}
+
+export async function runEngine() {
+  const config = await loadConfig()
+  const { cves } = config
+  if (!cves || !Array.isArray(cves)) {
+    throw new Error('Invalid config; check `cves` property.')
+  }
+
+  for (const cve of cves) {
+    await processCVE(cve)
+  }
+}
