+++ src/main.ts
@@ -1,36 +1,19 @@
-import { readFile } from "node:fs/promises"
-import { processCVE } from "./cve/processor"
+import { runEngine } from "./engine";
 
-interface CveBoxConfig {
-  cves: string[]
-}
-
-async function loadConfig() {
-  const f = process.env.CVEBOX_CONFIG_PATH || '/etc/cvebox/cvebox.config.json'
-  try {
-    const raw = await readFile(f, 'utf8')
-    const config: CveBoxConfig = JSON.parse(raw)
-    return config
-  } catch (err: any) {
-    if (err.code == 'ENOENT') {
-      throw new Error('Could not load config file: ' + f)
-    }
-    throw err
-  }
-}
 
 async function main() {
-  const config = await loadConfig()
-  const { cves } = config
-  if (!cves || !Array.isArray(cves)) {
-    throw new Error('Invalid config; check `cves` property.')
-  }
+  const [, , mode] = process.argv
+  switch (mode) {
+    case 'engine':
+      return runEngine()
 
-  for (const cve of cves) {
-    await processCVE(cve)
+    case 'scheduler':
+      throw new Error('Scheduler not implemented yet.')
+
+    default:
+      throw new Error(`Invalid mode (${mode || '<undefined>'}) – invoke via "node main.js <mode>`)
   }
 }
-
 
 main()
   .then(() => console.log(`[main] completed successfully`))
